Bots v0.1.0 Bots.BotsSupervisor

This supervisor for each bots in the list of specifications creates process according to bot’s type. If type is :active will be created process with long-polling implementation. If type is :passive will becreated process with webhook implementation.

Functions

init(bots_specs)

Callback implementation for c:Supervisor.init/1.

send_msg_to_worker(name, msg)

Searching bot’s process by name and puts into it message

start_link(bots_specs)